Update sign to support multi-chains
